% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{add_block}
\alias{add_block}
\title{Add a block to a stan model}
\usage{
add_block(x, name, lines, options)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{name}{name of the block}

\item{lines}{character vector, lines in the block}

\item{options}{any options to the block}
}
\value{
a \code{\link{stanedit}} object containing the new block
}
\description{
Add a block to a stan model. If that block exists, it will be removed first.
}
