% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barebones}
\alias{barebones}
\title{Strip model code to its bare bones}
\usage{
barebones(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector of the model code
}
\description{
Strip model code to its bare bones
}
\keyword{internal}
