% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check model}
\alias{check model}
\alias{check_model}
\title{Check model and produce messages/warnings if any problems}
\usage{
check_model(x)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}
}
\value{
invisible \code{x} (called for it's side effects)
}
\description{
This function currently only checks for balanced braces.
}
\seealso{
\code{\link{stanedit}}
}
\keyword{internal}
