% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{clean_model}
\alias{clean_model}
\title{Strip model code to its bare bones}
\usage{
clean_model(x)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}
}
\value{
the updated \code{stanedit} object
}
\description{
Cleans the model by working out correct indents, removing long comments and
merging lines (if desired)
}
\seealso{
\code{\link{stanedit}}
}
\keyword{internal}
