% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanedit.R
\name{Equals.stanedit}
\alias{Equals.stanedit}
\alias{==.stanedit}
\alias{`==.stanedit`}
\title{Check if two models are equal}
\usage{
\method{==}{stanedit}(e1, e2, ...)
}
\arguments{
\item{e1}{a \code{\link{stanedit}}}

\item{e2}{a \code{\link{stanedit}}}

\item{...}{ignored}
}
\value{
TRUE or FALSE, depending on whether the models are equal or not
}
\description{
Ignores differences in the model name.
}
\examples{
model_file_name <- system.file(package = "rbi", "PZ.bi")
PZ <- stanedit(filename = model_file_name)
PZ == PZ # TRUE
}
