% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Extract.stanedit}
\alias{Extract.stanedit}
\alias{[.stanedit}
\alias{`[.stanedit`}
\title{Subset model lines}
\usage{
\method{[}{stanedit}(x, i, ...)
}
\arguments{
\item{x}{A stanedit}

\item{i}{A vector of line numbers}

\item{...}{ignored}
}
\value{
a character string of the extracted model lines(s)
}
\description{
Extracts a subset of lines from the model.
}
\examples{
model_file_name <- system.file(package = "rbi", "PZ.bi")
PZ <- stanedit(filename = model_file_name)
PZ[3:4]
}
