% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Extract_assign.stanedit}
\alias{Extract_assign.stanedit}
\alias{[<-.stanedit}
\alias{`[<-.stanedit`}
\title{Subset and replace model lines}
\usage{
\method{[}{stanedit}(x, i, ...) <- value
}
\arguments{
\item{x}{A stanedit}

\item{i}{A vector of line numbers}

\item{...}{ignored}

\item{value}{A vector of the same length as \code{i}, containing the
replacement strings}
}
\value{
the updated \code{stanedit} object
}
\description{
Extracts a subset of lines from the model and assigns new character strings.
}
\examples{
model_file_name <- system.file(package = "rbi", "PZ.bi")
PZ <- stanedit(filename = model_file_name)
PZ[3:4] <- c("const e = 0.4", "const m_l = 0.05")
}
