% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{find_declaration}
\alias{find_declaration}
\title{Find the declaration of a variable in a stan model}
\usage{
find_declaration(x, name)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{name}{of the block to find}

\item{inner}{only return the inner part of the block (not the block
definition)}
}
\value{
an integer, the line numbes where the variable is declared; an empty
integer if the declaration is not found
}
\description{
Finds a block and returns the range of line numbers encompassed by that
block.
}
\seealso{
\code{\link{stanedit}}
}
\keyword{internal}
