% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_block}
\alias{get_block}
\title{Get the contents of a block in a stan model}
\usage{
get_block(x, name, shell = FALSE)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{name}{name of the block}

\item{shell}{if TRUE (default:FALSE), will return the shell (i.e., the
definition of the block) as well as content; this is useful, e.g., to see
options passed to a \code{transition} or \code{ode} block}
}
\value{
a character vector of the lines in the block
}
\description{
Returns the contents of a block in a stan model as a character vector of
lines.
}
\examples{
model_file_name <- system.file(package = "stanedit", "regression.stan")
reg <- stanedit(filename = model_file_name)
get_block(reg, "parameters")
}
