% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_vars}
\alias{get_vars}
\title{Extract the variables declared in a stan model}
\usage{
get_vars(x, block)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{block}{character (optional), the block or blocks in which to search;
by default the whole model will be searched}
}
\value{
a character vector of variable names
}
\description{
Extract the variables declared in a stan model
}
