% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.r
\name{move}
\alias{move}
\title{Move stan code elsewere}
\usage{
move(x, lines, ...)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{lines}{vector of line numbers to move}

\item{...}{arguments passed to \code{\link{insert_lines}}}
}
\value{
a character vector of variable names
}
\description{
Move stan code elsewere
}
\examples{
model_file_name <- system.file(package = "stanedit", "regression.stan")
reg <- stanedit(filename = model_file_name)
move(reg, find_declaration(reg, "alpha"), at_end_of = "data")
}
