% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.stanedit}
\alias{print.stanedit}
\title{Print the lines of a stan model}
\usage{
\method{print}{stanedit}(x, spaces = 2, screen = interactive(), ...)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{spaces}{number of spaces for indentation}

\item{screen}{whether to print to screen (default: TRUE). In that case, line
numbers will be added; otherwise, a character vector will be returned.}

\item{...}{ignored}
}
\value{
if \code{screen} is \code{FALSE}, a character vector of model lines
}
\description{
Prints all lines in a stan model
}
\examples{
model_file_name <- system.file(package = "stanedit", "regression.stan")
reg <- stanedit(filename = model_file_name)
print(reg, spaces = 4)
}
\keyword{internal}
