% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{remove_contained}
\alias{remove_contained}
\title{Remove everything between opening/closing characters, including nesting}
\usage{
remove_contained(x, opening, closing)
}
\arguments{
\item{x}{a character vector}

\item{opening}{opening character}

\item{closing}{closing character}
}
\value{
stanedit object with the bits removed
}
\description{
Useful for e.g. removing bounds, array indices etc.
}
