% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanedit.R
\name{stanedit}
\alias{stanedit}
\title{stanedit}
\usage{
stanedit(object, filename)
}
\arguments{
\item{object}{either a \code{stanmodel} object as generated by
\code{\link[rstan:stan_model]{rstan::stan_model()}} or a \code{CmdStanModel} object as generated by
\code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}, or a character vector representing model code.}

\item{filename}{name of a file containing model code, if \code{object} is not
given.}
}
\value{
A \code{{stanedit}} object containing clean model code. This will
have some formatting applied to it for further processing so line numbers
may not correspond to the original model, and comments will be removed.
}
\description{
\code{stanedit} prepares stan model code for programmatically editing.
}
\examples{
model_file_name <- system.file(package = "stanedit", "regression.stan")
reg <- stanedit(filename = model_file_name)
reg
}
