% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_model}
\alias{write_model}
\title{Writes a bi model to a file.}
\usage{
write_model(x, filename)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{filename}{name of the file to be written}
}
\value{
the return value of the \code{\link{writeLines}} call.
}
\description{
Writes a bi model to a file given by \code{filename}. The extension '.bi'
will be added if necessary.
}
\examples{
model_file_name <- system.file(package = "stanedit", "regression.stan")
reg <- stanedit(filename = model_file_name)
new_file_name <- tempfile("reg", fileext = ".stan")
write_model(reg, new_file_name)
}
\seealso{
\code{\link{stanedit}}
}
