% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{find_block}
\alias{find_block}
\title{Find a block in a stan model}
\usage{
find_block(x, name, inner = FALSE)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{name}{of the block to find}

\item{inner}{only return the inner part of the block (not the block
definition)}
}
\value{
an integerr vector, the range of line numbers
}
\description{
Finds a block and returns the range of line numbers encompassed by that
block.
}
\seealso{
\code{\link{stanedit}}
}
\keyword{internal}
