% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_declaration}
\alias{find_declaration}
\title{Find the declaration of a variable in a stan model}
\usage{
find_declaration(x, name)
}
\arguments{
\item{x}{a \code{\link{stanedit}} object}

\item{name}{of the block to find}
}
\value{
an integer, the line numbes where the variable is declared; an empty
integer if the declaration is not found
}
\description{
Finds a block and returns the range of line numbers encompassed by that
block.
}
\examples{
model_file_name <- system.file(package = "stanedit", "regression.stan")
reg <- stanedit(filename = model_file_name)
find_declaration(reg, "alpha")
}
\seealso{
\code{\link{stanedit}}
}
\keyword{internal}
